requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.4
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.35
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
